Display weather information in HTML format
This initial commit takes weather data from DarkSky and converts
it to a simple HTML message almost identical to the python
version.

Dark Sky is used because it contains tons of information that
other providers lack. Since we want as much information as
possible, Dark Sky is the best option currently.