Use lenses instead of derive generic for json
The old code was incredibly verbose, long, and annoying to extend.
This new code uses lenses which look very similar to the python code.
This makes the code look nicer and means that new things can be added
with only 1 new line of code instead of several.

As a result, all of the datatypes have been removed along with a couple
of helper functions.