Fail gracefully when the weather cannot be found
Previously, a failure in the weather file would crash the rest
of the program. Using a similar technique to XKCD, the program will
instead return an error String and still allow the rest of the
program to continue.

If one part of weather fails, everything in weather fails which
should probably be changed at some point.