Display APOD every day unless it fails
APOD provides a picture and a description every day. This program
is now capable of scraping them. RSS is not used because the rss
feed provides incredibly small images that are not of any use.

If APOD has a video instead of an image, it will gracefully fail
and return an empty string in place of real html.