Display SMBC comic every day
The SMBC comic is updated every day. It also doesn't mark when the
last update was so tracking this would have been very hard anyway.
The entire feed item is sent because it includes a large amount of
information and isn't work the effort needed to scrape away the
unimportant parts.