Fix Some String Errors
Remove `stringsearch` package.

Improve RSS feed so that it uses Text (already used in Text.Feed itself). Note
that this involves converting it to a Text.Lazy instead of a strict version.

Allow lenient decoding in `getHttp`. Previously, when it could not decode with
utf8, the function would give an exception and crash the program. Now, it will
try and replace with the utf8 character due to `lenientDecode`.

The problem originally arose since decodeUtf8 would freak out in the extended
ascii from ISO Latin. Although that sounds incredibly niche, APOD uses ISO
Latin, and it came up in the name of a photographer.

I also suggest that we try to streamline our system in general for dealing with
our various string types. Similarly, our use of `http-conduit` will likely kick
us in the butt in a few weeks. And of course, our web scraping could be
improved. The configuration could replace `read` with a parser or at least
`readMaybe`. And lastly, we could probably parallelize our system with
`parallel-io`, since we are essentially running a bunch of separate IO actions.

And to add the final cherry, we should probably put the above topics into the
TODO instead of my commit.